[7/n] use 'unsigned' for address
The original code used a 'size_t' variable (64-bit on modern platforms)
to store an address. That makes perfect sense.

In the context of openMSX, 'debuggables' only use a 32-bit address (more
than sufficient for any MSX device).

Printing 'size_t' variables via printf() is different in Linux and
Windows. The code used a macro to handle this difference.

Because of the combination of the above I decided to just use a 32-bit
variable to store addresses in thise code. And as a bonus that allowed
to remove some casts when interfacing with other openMSX code.

+++ src/imgui/ImGuiDebugger.cc
@@ -45,13 +45,13 @@ namespace openmsx {
 DebuggableEditor::DebuggableEditor(ImGuiManager& manager_)
 	: manager(&manager_)
 {
-	ReadFn = [](const ImU8* userdata, size_t offset) -> ImU8 {
+	ReadFn = [](const ImU8* userdata, unsigned offset) -> ImU8 {
 		auto* debuggable = reinterpret_cast<CallbackInfo*>(const_cast<ImU8*>(userdata))->debuggable;
-		return debuggable->read(narrow<unsigned>(offset));
+		return debuggable->read(offset);
 	};
-	WriteFn = [](ImU8* userdata, size_t offset, ImU8 data) -> void {
+	WriteFn = [](ImU8* userdata, unsigned offset, ImU8 data) -> void {
 		auto* debuggable = reinterpret_cast<CallbackInfo*>(userdata)->debuggable;
-		debuggable->write(narrow<unsigned>(offset), data);
+		debuggable->write(offset, data);
 	};
 }
